Refactored weather-mv to work with pluggable Data Sinks.
- Created abstraction for Data Sinks (where weather data can be moved to)
- Pulled out existing pipeline into a BigQuery sink
- Added test stubs
- Added `dataclasses` as a project dependency for older python versions

TODO:
- e2e test of the weather mover
- implement tests
- document
