+++ weather_mv/loader_pipeline/sinks_test.py
@@ -1,0 +1,62 @@
+# Copyright 2021 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import unittest
+from functools import wraps
+
+import weather_mv
+
+
+class TestDataBase(unittest.TestCase):
+    def setUp(self) -> None:
+        self.test_data_folder = f'{next(iter(weather_mv.__path__))}/test_data'
+
+
+def _handle_missing_grib_be(f):
+    @wraps(f)
+    def decorated(*args, **kwargs):
+        try:
+            return f(*args, **kwargs)
+        except ValueError as e:
+            # Some setups may not have Cfgrib installed properly. Ignore tests for these cases.
+            e_str = str(e)
+            if "Consider explicitly selecting one of the installed engines" not in e_str or "cfgrib" in e_str:
+                raise
+
+    return decorated
+
+
+class OpenDatasetTest(unittest.TestCase):
+
+    def test_local_copy_created(self):
+        pass
+
+    def test_opens_grib_files(self):
+        pass
+
+    def test_accepts_xarray_kwargs(self):
+        pass
+
+    def test_counts_success_metric(self):
+        pass
+
+    def test_counts_failure_metric(self):
+        pass
+
+    def test_opens_grib_in_memory(self):
+        pass
+
+
+if __name__ == '__main__':
+    unittest.main()
